Event Flow Refactoring
