+++ lib/event_flow/ui/event_flow_section.dart
@@ -1,14 +1,13 @@
+import 'package:flutter/material.dart';
+
 import 'package:festival_flutterturkiye_org/core/model/session.dart';
 import 'package:festival_flutterturkiye_org/core/model/speaker.dart';
 import 'package:festival_flutterturkiye_org/core/ui/responsive_builder.dart';
-import 'package:festival_flutterturkiye_org/widgets/event_flow_section/event_flow_session_point.dart';
-import 'package:festival_flutterturkiye_org/widgets/event_flow_section/session_info_field.dart';
-import 'package:festival_flutterturkiye_org/widgets/event_flow_section/session_time_field.dart';
-import 'package:flutter/material.dart';
-
 import 'package:festival_flutterturkiye_org/core/ui/section_subtitle.dart';
 import 'package:festival_flutterturkiye_org/core/ui/section_title.dart';
 import 'package:festival_flutterturkiye_org/core/utils/theme_helper.dart';
+import 'package:festival_flutterturkiye_org/event_flow/ui/session_info_field.dart';
+import 'package:festival_flutterturkiye_org/event_flow/ui/session_time_field.dart';
 
 class EventFlowSection extends StatelessWidget {
   const EventFlowSection({Key key}) : super(key: key);
@@ -64,26 +63,25 @@ class _SessionWidget extends StatelessWidget {
     final startingTime = _getTime(session.startingTime);
     final dueTime = _getTime(session.startingTime.add(session.duration));
 
-    return ResponsiveBuilder(
-      smallWidget: Padding(
-        padding: const EdgeInsets.only(bottom: 16),
-        child: SessionInfoField(
+    return Padding(
+      padding: const EdgeInsets.only(bottom: 16),
+      child: ResponsiveBuilder(
+        smallWidget: _SmallSessionWidget(
           session: session,
           speaker: speaker,
-          isSmallScreen: true,
         ),
-      ),
-      mediumWidget: _LargeSessionWidget(
-        startingTime: startingTime,
-        dueTime: dueTime,
-        session: session,
-        speaker: speaker,
-      ),
-      largeWidget: _LargeSessionWidget(
-        startingTime: startingTime,
-        dueTime: dueTime,
-        session: session,
-        speaker: speaker,
+        mediumWidget: _LargeSessionWidget(
+          startingTime: startingTime,
+          dueTime: dueTime,
+          session: session,
+          speaker: speaker,
+        ),
+        largeWidget: _LargeSessionWidget(
+          startingTime: startingTime,
+          dueTime: dueTime,
+          session: session,
+          speaker: speaker,
+        ),
       ),
     );
   }
@@ -111,28 +109,90 @@ class _LargeSessionWidget extends StatelessWidget {
   final Speaker speaker;
 
   @override
-  Widget build(BuildContext context) => Padding(
-        padding: const EdgeInsets.only(bottom: 16),
-        child: Row(
-          crossAxisAlignment: CrossAxisAlignment.start,
-          children: [
-            Expanded(
-              child: EventFlowSessionText(
-                text: '$startingTime - $dueTime',
-                sessionStatus: session.status,
-                textAlign: TextAlign.end,
-              ),
+  Widget build(BuildContext context) => Row(
+        crossAxisAlignment: CrossAxisAlignment.start,
+        children: [
+          Expanded(
+            child: EventFlowSessionText(
+              text: '$startingTime - $dueTime',
+              sessionStatus: session.status,
+              textAlign: TextAlign.end,
             ),
-            EventFlowSessionPoint(sessionStatus: session.status),
-            Expanded(
-              child: SessionInfoField(
-                session: session,
-                speaker: speaker,
-              ),
+          ),
+          _EventFlowSessionPoint(sessionStatus: session.status),
+          Expanded(
+            child: SessionInfoField(
+              session: session,
+              speaker: speaker,
             ),
-          ],
-        ),
+          ),
+        ],
       );
+}
+
+class _SmallSessionWidget extends StatelessWidget {
+  const _SmallSessionWidget({
+    @required this.session,
+    this.speaker,
+    Key key,
+  })  : assert(session != null),
+        super(key: key);
+
+  final Session session;
+  final Speaker speaker;
+
+  @override
+  Widget build(BuildContext context) => Row(
+        mainAxisSize: MainAxisSize.min,
+        crossAxisAlignment: CrossAxisAlignment.start,
+        children: [
+          _EventFlowSessionPoint(sessionStatus: session.status),
+          Expanded(
+            child: SessionInfoField(
+              session: session,
+              speaker: speaker,
+              isSmallScreen: true,
+            ),
+          ),
+        ],
+      );
+}
+
+class _EventFlowSessionPoint extends StatelessWidget {
+  const _EventFlowSessionPoint({
+    @required this.sessionStatus,
+    Key key,
+  })  : assert(sessionStatus != null),
+        super(key: key);
+  final SessionStatus sessionStatus;
+
+  @override
+  Widget build(BuildContext context) {
+    Color pointColor;
+    double radius;
+
+    switch (sessionStatus) {
+      case SessionStatus.active:
+        pointColor = ThemeHelper.eventPointColor;
+        radius = 28.0;
+        break;
+      case SessionStatus.passed:
+        pointColor = ThemeHelper.appBarActionColor;
+        radius = 20.0;
+        break;
+      case SessionStatus.waiting:
+      default:
+        pointColor = ThemeHelper.blueColor;
+        radius = 20.0;
+        break;
+    }
+    return Container(
+      height: radius,
+      width: radius,
+      margin: const EdgeInsets.symmetric(horizontal: 12),
+      decoration: BoxDecoration(shape: BoxShape.circle, color: pointColor),
+    );
+  }
 }
 
 const _speakers = <Speaker>[
